backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Mako==1.3.6
MarkupSafe==3.0.2
motor==3.6.0
packaging==24.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
